feat: Divide mock data by apis
