+++ src/mocks/browser.js
@@ -1,9 +1,9 @@
 import { setupWorker } from "msw";
-import { userHandlers } from "./userHandler";
+import { chatHandlers } from "./chatHandler";
+import { paymentHandlers } from "./paymentHandler";
 import { portfolioHandlers } from "./portfolioHandler";
 import { quotationHandlers } from "./quotationHandler";
-import { paymentHandlers } from "./paymentHandler";
-import { matchHandlers } from "./matchHandler";
+import { userHandlers } from "./userHandler";
 
 // This configures a Service Worker with the given request handlers.
 export const worker = setupWorker(
@@ -11,5 +11,5 @@ export const worker = setupWorker(
   ...portfolioHandlers,
   ...quotationHandlers,
   ...paymentHandlers,
-  ...matchHandlers,
+  ...chatHandlers,
 );
