+++ src/mocks/chatData.js
@@ -1,0 +1,7 @@
+export const chatId = {
+  success: true,
+  response: {
+    chatId: 999,
+  },
+  error: null,
+};
