+++ src/mocks/commonData.js
@@ -1,0 +1,5 @@
+export const sucess = {
+  success: true,
+  response: null,
+  error: null,
+};
