--- src/mocks/matchHandler.js
@@ -1,25 +1,0 @@
-import { rest } from "msw";
-import { chatId } from "./responseData";
-
-async function sleep(ms) {
-  return new Promise((resolve) => {
-    setTimeout(resolve, ms);
-  });
-}
-
-export const matchHandlers = [
-  rest.post("/match/chat", async (req, res, ctx) => {
-    await sleep(500);
-    const isAuthenticated = localStorage.getItem("token");
-    if (!isAuthenticated) {
-      return res(
-        ctx.status(403),
-        ctx.json({
-          code: 403,
-          message: "Not authorized",
-        }),
-      );
-    }
-    return res(ctx.status(200), ctx.json(chatId));
-  }),
-];
