+++ src/mocks/paymentsData.js
@@ -1,0 +1,9 @@
+// /payments/save
+// successResponse
+
+// /payments/confirm
+export const paymentsConfirm = {
+  success: true,
+  response: "success", // 실패하면 fail
+  error: null,
+};
