+++ src/mocks/quotationData.js
@@ -1,0 +1,36 @@
+export const quotationList = {
+  success: true,
+  response: {
+    status: "완료", // 미완료
+    quotations: [
+      {
+        id: 1,
+        title: "헤어",
+        price: 350000,
+        company: "***",
+        description: "",
+        status: "완료",
+        modifiedAt: "2023.09.13",
+      },
+      {
+        id: 2,
+        title: "헤어",
+        price: 350000,
+        company: "***",
+        description: "",
+        status: "완료",
+        modifiedAt: "2023.09.13",
+      },
+      {
+        id: 3,
+        title: "헤어",
+        price: 350000,
+        company: "***",
+        description: "",
+        status: "미완료",
+        modifiedAt: "2023.09.13",
+      },
+    ],
+  },
+  error: null,
+};
