+++ src/mocks/userData.js
@@ -1,0 +1,28 @@
+// /user/signup
+export const signupResponse = {
+  success: true,
+  response: {
+    userId: 1,
+  },
+  error: null,
+};
+
+// /user/login
+// SuccessResponse
+
+// /user
+// SuccessResponse
+
+// /user/info
+export const infoResponse = {
+  success: true,
+  response: {
+    userId: 1,
+    username: "김하나",
+    email: "hana0101@naver.com",
+    role: "couple",
+    grade: "premium",
+    payedAt: "2023년 09월 15일",
+  },
+  error: null,
+};
